애니메이션 수정

+++ Assets/02_Scripts/02_Character/Enemy.cs
@@ -5,9 +5,9 @@ using UnityEngine;
 
 public class Enemy : MonoBehaviour
 {
-    [SerializeField] private BigInteger hp = 500;
+    [SerializeField] private float hp = 500;
     
-    public void GetDamage(BigInteger damage)
+    public void GetDamage(float damage)
     {
         FloatingTextController.I.CreateFloatingText(damage.ToString(), transform);
         hp -= damage;
